:lipstick: Style: 시설 상세 컴포넌트 분리
